+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   That's why we're going to offer **two** releases for an indefinite period of time
   - `3.0.0` with a npm tag `latest`: Vue 3.x
   - `3.0.0-vue2` with a npm tag `vue2`: Vue 2.x
+- The class name of the `$gtm` instance changed from `AnalyticsPlugin` to `VueGtmPlugin`
 
 ## Other Changes